Create playRound function, Debug getHumanChoice
added a function to play one round of rock paper
scissors. if both choices are the same return a
log claiming a draw. if human wins return win
message and increase humanScore. Else, increase
computerScore and return log of lose message.
Removed variable rock, paper, and scissors. create
variable choice within getHumanChoice function and
assign it with the answer to the prompt.